[Auto Generated] v3.1.0-beta.7

+++ Sources/code/application/Models/Cart/ProductAvailabilityCartAppModel.swift
@@ -16,7 +16,7 @@ public extension ApplicationClient.Cart {
         
         public var isValid: Bool?
         
-        public var otherStoreQuantity: Int?
+        public var otherStoreQuantity: Int64?
         
         public var sizes: [String]?
         
@@ -37,7 +37,7 @@ public extension ApplicationClient.Cart {
             
         }
 
-        public init(availableSizes: [ProductAvailabilitySize]? = nil, deliverable: Bool? = nil, isValid: Bool? = nil, otherStoreQuantity: Int? = nil, outOfStock: Bool? = nil, sizes: [String]? = nil) {
+        public init(availableSizes: [ProductAvailabilitySize]? = nil, deliverable: Bool? = nil, isValid: Bool? = nil, otherStoreQuantity: Int64? = nil, outOfStock: Bool? = nil, sizes: [String]? = nil) {
             
             self.outOfStock = outOfStock
             
@@ -106,7 +106,7 @@ public extension ApplicationClient.Cart {
             
             
             do {
-                otherStoreQuantity = try container.decode(Int.self, forKey: .otherStoreQuantity)
+                otherStoreQuantity = try container.decode(Int64.self, forKey: .otherStoreQuantity)
             
             } catch DecodingError.typeMismatch(let type, let context) {
                 print("Type '\(type)' mismatch:", context.debugDescription)
